snapdict-core/requirements.txt
fastapi>=0.100.0
uvicorn>=0.15.0
python-multipart
pytesseract>=0.3.10
Pillow>=10.0.0
nltk>=3.8.1
pytest
opencv-python>=4.8.0
numpy>=1.24.0
pydantic>=2.0.0
aiofiles
python-dotenv
